feature: downgrade "json-schema-validator" dependency
Josejulio realized that there is an issue with the
"json-schema-validator"[1] dependency we depend on. Basically the custom
formatters are not being honored, and that causes some date times to not
be read as expected.

Josejulio is already working on a fix[2], but in the meantime we will
downgrade the library we depend on to a previous version.

[1]: https://github.com/networknt/json-schema-validator/issues/784
[2]: https://github.com/networknt/json-schema-validator/pull/787